Move AttendanceStatus
