+++ app/components/UserAttendance/index.js
@@ -1,0 +1,2 @@
+export { default as AttendanceStatus } from './AttendanceStatus';
+export { default as AttendanceModal } from './AttendanceModal';
